Refactored resume fixtures for use by test_chat.py
